Update styles from the app.
Trying to make it a little more streamlined to separate styles that are
just for the marketing site so I can just dump in files from the app or
pattern library without having to think about it too much.

Why I decided that having essentially the same set of files living in
three repositories was a good idea, I don’t know.